requirements.txt
azure-functions
tensorflow
textgenrnn